Some code refratoring
