+++ assets/js/app/tree_view/components/Header.js
@@ -1,33 +1,24 @@
-import React from 'react';
+import React, {useState} from "react";
 import PropTypes from 'prop-types';
 
-import { Div } from 'react-treebeard';
-
-// Example: Customising The Header Decorator To Include Icons
-const Header = ({onSelect, style, customStyles, node}) => {
-    const iconType = node.children ? 'folder' : 'file-text';
-    const iconClass = `fa fa-${iconType}`;
+const Header = ({ node, style, iconClass }) => {
+    const [iconTypeClass, setIconTypeClass] = useState(iconClass);
     const iconStyle = {marginRight: '5px'};
-
     return (
-        <div style={style.base} onClick={onSelect}>
-            <Div style={node.selected ? {...style.title, ...customStyles.header.title} : style.title}>
-                <i className={iconClass} style={iconStyle}/>
-                {node.name}
-            </Div>
+      <div style={style.base}>
+        <div style={{ ...style.title }}>
+            <i className={`fa fa-${iconTypeClass}`} style={iconStyle}/>
+            {node.name}
         </div>
+      </div>
     );
 };
 
+
 Header.propTypes = {
-    onSelect: PropTypes.func,
     node: PropTypes.object,
     style: PropTypes.object,
-    customStyles: PropTypes.object
-};
-
-Header.defaultProps = {
-    customStyles: {}
+    iconClass: PropTypes.string
 };
 
 export default Header;
